+++ Controller/Admin/Users/ControllerUsers.php
@@ -69,8 +69,7 @@ class ControllerUsers extends Controller
     // Modification d'un user :
     public function updateuser()
     {
-        $id_user = $this->request->getParameter("id");
-
+        $id_user = $this->request->getParameter("id");      
         if (isset($_POST["update"]) && !empty($_POST["firstname"]) && !empty($_POST["name"])) {
             $errors                = array();
             $messages              = array();
@@ -84,9 +83,6 @@ class ControllerUsers extends Controller
             $twitter               = $_POST["twitter"];
             $website               = $_POST["website"];
             $date_birth            = $_POST["date_birth"];
-            $fileinfo              = @getimagesize($_FILES["avatar"]["tmp_name"]);
-            $width                 = $fileinfo[0];
-            $height                = $fileinfo[1];
             $extensions_authorized = array(
                 "gif",
                 "png",
@@ -94,41 +90,44 @@ class ControllerUsers extends Controller
                 "jpeg"
             );
             $extension_upload      = pathinfo($_FILES["avatar"]["name"], PATHINFO_EXTENSION);
-            $time                  = date("Y-m-d-H-i-s");
-            $avatarname            = str_replace(' ', '-', strtolower($_FILES['avatar']['name']));
-            $avatarname            = preg_replace("/\.[^.\s]{3,4}$/", "", $avatarname);
-            $avatarname            = "{$time}-{$id_user}-avatar.{$extension_upload}";
-            $destination           = ROOT_PATH . 'public/images/avatars';
-
-            if (!file_exists($_FILES["avatar"]["tmp_name"])) {
-                $this->user->changeUserFromAdmin($id_user, $status, $firstname, $name, $email, $city, $linkedin, $github,
-                $twitter, $website, $date_birth);
-                $this->message->userAccountUpdated($id_user);
-            } else if (!in_array($extension_upload, $extensions_authorized)) {
+            if (!in_array($extension_upload, $extensions_authorized)) {
                 $errors['errors'] = 'L\'extension du fichier n\'est pas autorisée.';
                 if (!empty($errors)) {
                     $_SESSION['errors'] = $errors;
                     header('Location: ' . BASE_ADMIN_URL . 'users/userread/' . $id_user);
                     exit;
                 }
-            } else if (($_FILES["avatar"]["size"] > 1000000)) {
-                $errors['errors'] = 'Le fichier est trop lourd.';
-                if (!empty($errors)) {
-                    $_SESSION['errors'] = $errors;
-                    header('Location: ' . BASE_ADMIN_URL . 'users/userread/' . $id_user);
-                    exit;
-                }
-            } else if ($width < "800" && $height < "600") {
-                $errors['errors'] = 'Les dimensions sont trop petites. <br>Minimum : 800 X 600 px';
-                if (!empty($errors)) {
-                    $_SESSION['errors'] = $errors;
-                    header('Location: ' . BASE_ADMIN_URL . 'users/userread/' . $id_user);
-                    exit;
+            }
+            if (!empty($_FILES['avatar']['tmp_name']) && file_exists($_FILES['avatar']['tmp_name'])) {
+                $fileinfo    = @getimagesize($_FILES["avatar"]["tmp_name"]);
+                $width       = $fileinfo[0];
+                $height      = $fileinfo[1];
+                $time        = date("Y-m-d-H-i-s");
+                $avatarname  = str_replace(' ', '-', strtolower($_FILES['avatar']['name']));
+                $avatarname  = preg_replace("/\.[^.\s]{3,4}$/", "", $avatarname);
+                $avatarname  = "{$time}-{$id_user}-avatar.{$extension_upload}";
+                $destination = ROOT_PATH . 'public/images/avatars';
+                if (($_FILES["avatar"]["size"] > 1000000)) {
+                    $errors['errors'] = 'Le fichier est trop lourd.';
+                    if (!empty($errors)) {
+                        $_SESSION['errors'] = $errors;
+                        header('Location: ' . BASE_ADMIN_URL . 'users/userread/' . $id_user);
+                        exit;
+                    }
+                } else if ($width < "800" && $height < "600") {
+                    $errors['errors'] = 'Les dimensions sont trop petites. <br>Minimum : 800 X 600 px';
+                    if (!empty($errors)) {
+                        $_SESSION['errors'] = $errors;
+                        header('Location: ' . BASE_ADMIN_URL . 'users/userread/' . $id_user);
+                        exit;
+                    }
+                } else {
+                    move_uploaded_file($_FILES['avatar']['tmp_name'], $destination . "/" . $avatarname);
+                    $this->user->changeUserImageFromAdmin($id_user, $status, $firstname, $name, $avatarname, $email, $city, $linkedin, $github, $twitter, $website, $date_birth);
+                    $this->message->userAccountUpdated($id_user);
                 }
             } else {
-                move_uploaded_file($_FILES['avatar']['tmp_name'], $destination . "/" . $avatarname);
-                $this->user->changeUserImageFromAdmin($id_user, $status, $firstname, $name, $avatarname, $email, $city, $linkedin, $github,
-                $twitter, $website, $date_birth);
+                $this->user->changeUserFromAdmin($id_user, $status, $firstname, $name, $email, $city, $linkedin, $github, $twitter, $website, $date_birth);
                 $this->message->userAccountUpdated($id_user);
             }
         } else {
@@ -138,7 +137,6 @@ class ControllerUsers extends Controller
                 header('Location: ' . BASE_ADMIN_URL . 'users/userread/' . $id_user);
                 exit;
             }
-
         }
     }
 
@@ -199,6 +197,5 @@ class ControllerUsers extends Controller
         $this->user->restoreUser($id_user);
         $this->message->userRestored();
     }
-
 
 }
